Streaming: directory casing rename part 2
