+++ libraries/botframework-streaming/src/interfaces/IAssemblerParams.ts
