+++ libraries/botframework-streaming/src/interfaces/IHeader.ts
