+++ libraries/botframework-streaming/src/interfaces/IHttpContentHeaders.ts
