+++ libraries/botframework-streaming/src/interfaces/IReceiveRequest.ts
