+++ libraries/botframework-streaming/src/interfaces/IReceiveResponse.ts
