+++ libraries/botframework-streaming/src/interfaces/IResponsePayload.ts
