+++ libraries/botframework-streaming/src/interfaces/ISendPacket.ts
