+++ libraries/botframework-streaming/src/interfaces/ISocket.ts
