+++ libraries/botframework-streaming/src/interfaces/IStreamDescription.ts
