+++ libraries/botframework-streaming/src/interfaces/IStreamWrapper.ts
