+++ libraries/botframework-streaming/src/interfaces/IStreamingTransportClient.ts
