+++ libraries/botframework-streaming/src/interfaces/IStreamingTransportServer.ts
