+++ libraries/botframework-streaming/src/interfaces/ITransport.ts
