+++ libraries/botframework-streaming/src/interfaces/ITransportReceiver.ts
