+++ libraries/botframework-streaming/src/interfaces/ITransportSender.ts
