+++ libraries/botframework-streaming/src/interfaces/index.ts
