+++ libraries/botframework-streaming/src/namedPipe/NamedPipeClient.ts
