+++ libraries/botframework-streaming/src/namedPipe/NamedPipeServer.ts
