+++ libraries/botframework-streaming/src/namedPipe/NamedPipeTransport.ts
