+++ libraries/botframework-streaming/src/namedPipe/index.ts
