+++ libraries/botframework-streaming/src/payloadTransport/index.ts
