+++ libraries/botframework-streaming/src/payloads3/HeaderSerializer.ts
