+++ libraries/botframework-streaming/src/payloads3/SendOperations.ts
