+++ libraries/botframework-streaming/src/utilities/index.ts
