+++ libraries/botframework-streaming/src/utilities/protocol-base.ts
