+++ libraries/botframework-streaming/src/webSocket/BrowserWebSocket.ts
