+++ libraries/botframework-streaming/src/webSocket/WebSocketClient.ts
